+++ codestacktrace-api/src/main/java/com/biagiolibe/dev/codestacktrace/api/CodestacktraceApplication.java
@@ -1,4 +1,4 @@
-package com.biagiolibe.dev.codestacktrace;
+package com.biagiolibe.dev.codestacktrace.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
