+++ codestacktrace-api/src/main/java/com/biagiolibe/dev/codestacktrace/api/compiler/model/CompilationResponse.java
@@ -1,4 +1,4 @@
-package com.biagiolibe.dev.codestacktrace.compiler.model;
+package com.biagiolibe.dev.codestacktrace.api.compiler.model;
 
 import javax.tools.DiagnosticCollector;
 import javax.tools.JavaFileObject;
