+++ codestacktrace-api/src/main/java/com/biagiolibe/dev/codestacktrace/api/printer/ConsoleOutputCapturer.java
@@ -1,4 +1,4 @@
-package com.biagiolibe.dev.codestacktrace.printer;
+package com.biagiolibe.dev.codestacktrace.api.printer;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
